requirements.txt
streamlit
streamlit-analytics2
requests
urllib3
xlsxwriter
pandas
python-pptx
openpyxl